ECS - Changed ColliderComponent radius to float and added NetworkReconciliationSystem class.

+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/ColliderComponent.java
@@ -3,5 +3,5 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class ColliderComponent implements Component {
-    public double radius;
+    public float radius;
 }
